:truck: separate view tests into several files
